pyproject.toml
[project]
name = "shipr"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Paul-Sees-Solutions", email = "79617825+Paul-Sees-Solutions@users.noreply.github.com" }
]
dependencies = [
    "fastui",
    "pydantic",
    "aiohttp",
    "python-dotenv",
    "zeep",
    "datamodel-code-generator",
]

readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "pyinstaller",
]

[tool.hatch.metadata]
allow-direct-references = true
